+++ src/main/java/org/truffleruby/language/exceptions/TopLevelRaiseHandler.java
@@ -15,6 +15,7 @@ import org.truffleruby.core.kernel.AtExitManager;
 import org.truffleruby.language.RubyNode;
 import org.truffleruby.language.control.ExitException;
 import org.truffleruby.language.control.RaiseException;
+import org.truffleruby.language.dispatch.CallDispatchHeadNode;
 import org.truffleruby.language.objects.ReadObjectFieldNodeGen;
 
 import com.oracle.truffle.api.CompilerDirectives;
@@ -34,11 +35,13 @@ public class TopLevelRaiseHandler extends RubyNode {
     @Override
     public Object execute(VirtualFrame frame) {
         int exitCode = 0;
+        DynamicObject caughtException = null;
 
         try {
             body.execute(frame);
         } catch (RaiseException e) {
             DynamicObject rubyException = AtExitManager.handleAtExitException(getContext(), e);
+            caughtException = rubyException;
             setLastException(frame, rubyException);
             exitCode = statusFromException(rubyException);
         } catch (ExitException e) {
@@ -48,6 +51,10 @@ public class TopLevelRaiseHandler extends RubyNode {
 
             if (atExitException != null) {
                 exitCode = statusFromException(atExitException);
+            }
+
+            if (caughtException != null) {
+                handleSignalException(caughtException);
             }
         }
 
@@ -78,6 +85,13 @@ public class TopLevelRaiseHandler extends RubyNode {
         }
 
         setExceptionVariableNode.setLastException(frame, exception);
+    }
+
+    private void handleSignalException(DynamicObject exception) {
+        if (Layouts.BASIC_OBJECT.getLogicalClass(exception) == coreLibrary().getSignalExceptionClass()) {
+            // Calls raise(3) or no-op
+            CallDispatchHeadNode.getUncached().call(exception, "reached_top_level");
+        }
     }
 
 }