+++ src/main/ruby/truffleruby/core/exception.rb
@@ -469,6 +469,22 @@ class SignalException < Exception
     @signo = signo
     super(name_with_prefix)
   end
+
+  private def reached_top_level
+    if @signo == Signal::Names['VTALRM']
+      warn 'not acting on top level SignalException for SIGVTALRM as it is VM reserved'
+      return
+    end
+
+    begin
+      Signal.trap(@signo, 'SYSTEM_DEFAULT')
+    rescue ArgumentError
+      # some signals are reserved but we can raise them anyways
+      nil
+    end
+    Truffle::POSIX.raise_signal(@signo)
+  end
+
 end
 
 class StopIteration